[FIX Rename]Садиков Иван. Задача 2. Вариант 5. Умножение разреженных матриц (CCS)  (#507)
Переименовал все отключенные директории, синхронизировал include
