+++ tasks/omp/sadikov_I_SparseMatrixMultiplication/include/SparseMatrix.hpp
