+++ tasks/omp/sadikov_I_SparseMatrixMultiplication/include/ops_omp.hpp
@@ -4,7 +4,7 @@
 #include <utility>
 
 #include "core/task/include/task.hpp"
-#include "omp/Sadikov_I_SparseMatrixMultiplication_OMP/include/SparseMatrix.hpp"
+#include "omp/sadikov_I_SparseMatrixMultiplication/include/SparseMatrix.hpp"
 
 namespace sadikov_i_sparse_matrix_multiplication_task_omp {
 
@@ -14,7 +14,6 @@ class CCSMatrixOMP : public ppc::core::Task {
   SparseMatrix m_answerMatrix_;
 
  public:
-  // restart tests
   explicit CCSMatrixOMP(ppc::core::TaskDataPtr task_data) : Task(std::move(task_data)) {}
   bool PreProcessingImpl() override;
   bool ValidationImpl() override;
