+++ tasks/omp/sadikov_I_SparseMatrixMultiplication/perf_tests/main.cpp
@@ -9,11 +9,10 @@
 
 #include "core/perf/include/perf.hpp"
 #include "core/task/include/task.hpp"
-#include "omp/Sadikov_I_SparseMatrixMultiplication_OMP/include/SparseMatrix.hpp"
-#include "omp/Sadikov_I_SparseMatrixMultiplication_OMP/include/ops_omp.hpp"
+#include "omp/sadikov_I_SparseMatrixMultiplication/include/SparseMatrix.hpp"
+#include "omp/sadikov_I_SparseMatrixMultiplication/include/ops_omp.hpp"
 
 namespace {
-// restart tests
 std::vector<double> GetRandomMatrix(int size) {
   std::vector<double> data(size);
   std::random_device dev;
