+++ tasks/omp/sadikov_I_SparseMatrixMultiplication/src/ops_omp.cpp
@@ -1,9 +1,9 @@
-#include "omp/Sadikov_I_SparseMatrixMultiplication_OMP/include/ops_omp.hpp"
+#include "omp/sadikov_I_SparseMatrixMultiplication/include/ops_omp.hpp"
 
 #include <cstddef>
 #include <vector>
 
-#include "omp/Sadikov_I_SparseMatrixMultiplication_OMP/include/SparseMatrix.hpp"
+#include "omp/sadikov_I_SparseMatrixMultiplication/include/SparseMatrix.hpp"
 
 bool sadikov_i_sparse_matrix_multiplication_task_omp::CCSMatrixOMP::PreProcessingImpl() {
   auto fmatrix_rows_count = static_cast<int>(task_data->inputs_count[0]);
