+++ tasks/tbb/sadikov_I_SparseMatrixMultiplication/src/ops_tbb.cpp
@@ -1,9 +1,9 @@
-#include "tbb/sadikov_I_SparseMatMul_TBB/include/ops_tbb.hpp"
+#include "tbb/sadikov_I_SparseMatrixMultiplication/include/ops_tbb.hpp"
 
 #include <cstddef>
 #include <vector>
 
-#include "tbb/sadikov_I_SparseMatMul_TBB/include/SparseMatrix.hpp"
+#include "tbb/sadikov_I_SparseMatrixMultiplication/include/SparseMatrix.hpp"
 
 bool sadikov_i_sparse_matrix_multiplication_task_tbb::CCSMatrixTBB::PreProcessingImpl() {
   auto fmatrix_rows_count = static_cast<int>(task_data->inputs_count[0]);
